Create a base nbconvert class for apps that use nbconvert
